replace PyYAML with ruamel.yaml and add source info to --debug
This patch replaces PyYAML with ruamel.yaml for YAML parsing, and has
a new implementation of the configuration tree that keeps track of the
places where config values came from (i.e. YAML files or code). This
is probably bear-slow, but that doesn't hurt for this use case and
helps a lot letting users understand what's going on.