feature: production rules for preconditions
spinfile can now define "preconditions" for certain (sub-)
commands. This works by attaching a 'needs' key to a command name,
where command names are either 'name' for top-level commands, or
'group name' for second level commands. 'needs' can have a string of a
list value. spin assumes those precondition to be files (or
directories), and looks them up in a section called 'howtobuild' (for
now). For each target, that section has a production rule, that
consists of either a list of shell commands or spin sub-commands.