java: handle conflicting JDK installs
in case there already is a JDK installation, but it is not yet
memoized, unpacking the freshly downloaded new distribution into the
same directory will fail, since the JDK archives have r/o files. But
we cannot remove the installation directory beforehand, since we do
not know it's name: `java_home` is computed from the archive contents.